+++ src/decibels.rs
@@ -1,5 +1,5 @@
 use std::fmt::{Display, Formatter};
-use std::ops::{Add, Sub};
+use std::ops::{Add, Neg, Sub};
 
 /// Represents a Decibel-valued sound level
 #[derive(Copy, Clone, Debug)]
@@ -32,10 +32,29 @@ impl Display for Decibels {
     }
 }
 
+impl PartialEq for Decibels {
+    fn eq(&self, other: &Decibels) -> bool { self.inner == other.inner }
+}
+
+impl PartialOrd for Decibels {
+    fn partial_cmp(&self, other: &Decibels) -> Option<std::cmp::Ordering> { self.inner.partial_cmp(&other.inner) }
+}
+
+impl Decibels {
+    #[must_use]
+    pub fn abs(self) -> Decibels { Decibels { inner: self.inner.abs() } }
+}
+
 impl Sub for Decibels {
     type Output = Decibels;
 
     fn sub(self, other: Decibels) -> Decibels { Decibels { inner: self.inner - other.inner } }
+}
+
+impl Neg for Decibels {
+    type Output = Decibels;
+
+    fn neg(self) -> Decibels { Decibels { inner: -self.inner } }
 }
 
 impl Add for Decibels {